trainingTF/requirements.txt
tensorflow-cpu<2
numpy
pandas
sklearn
matplotlib
h5py